# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AIOps Monitoring Middleware API
  version: 1.5.0
  description: Provider-neutral monitoring API plus authenticated site-agent snapshot and alert ingestion. Implementations must not expose LibreNMS payloads, tokens, database access, or renderer-specific options.
servers:
  - url: http://127.0.0.1:4310
paths:
...
  /api/v1/monitoring/overview:
    get:
      operationId: getMonitoringOverview
      responses:
        '200':
          description: Normalized device and alert summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonitoringOverview' }
        '502': { $ref: '#/components/responses/UpstreamError' }
...
  /api/v1/monitoring/alerts:
    get:
      operationId: listMonitoringAlerts
      parameters:
        - in: query
          name: state
          schema: { type: string, enum: [all, active, ack, unack, open], default: all }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        '200':
          description: Normalized alerts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonitoringAlertList' }
  /api/v1/monitoring/alert-history:
    get:
      operationId: listMonitoringAlertHistory
      summary: Query a read-only LibreNMS alert log window of at most 31 days
      parameters:
        - { in: query, name: from, required: true, schema: { type: string, format: date-time } }
        - { in: query, name: to, required: true, schema: { type: string, format: date-time } }
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
        - { in: query, name: deviceId, schema: { type: string, maxLength: 160 } }
      responses:
        '200':
          description: Paged normalized alert history without sensitive LibreNMS fields
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonitoringAlertHistoryPage' }
        '400': { description: Invalid query or range longer than 31 days }
  /api/v1/monitoring/series/port-traffic:
    get:
      operationId: queryPortTrafficSeries
      parameters:
        - in: query
          name: deviceId
          required: true
          schema: { type: string, maxLength: 160 }
        - in: query
          name: portId
          required: true
          schema: { type: string, maxLength: 160 }
      responses:
        '200':
          description: Renderer-neutral rolling traffic series
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TimeSeriesFrame' }
        '404': { description: Port not found }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    UpstreamError:
      description: Provider unavailable; upstream response and credentials are never returned
  schemas:
    ProviderDescriptor:
      type: object
      additionalProperties: false
      required: [id, name, configured]
      properties:
        id: { type: string }
        name: { type: string }
        configured: { type: boolean }
...
    MonitoringOverview:
      type: object
      additionalProperties: false
      required: [generatedAt, provider, devices, alerts]
      properties:
        generatedAt: { type: string, format: date-time }
        provider: { $ref: '#/components/schemas/ProviderDescriptor' }
        devices:
          type: object
          additionalProperties: false
          required: [total, up, down, states]
          properties:
            total: { type: integer, minimum: 0 }
            up: { type: integer, minimum: 0 }
            down: { type: integer, minimum: 0 }
            states:
              type: object
              additionalProperties: { type: integer, minimum: 0 }
        alerts:
          type: object
          additionalProperties: false
          required: [total, active, severities]
          properties:
            total: { type: integer, minimum: 0 }
            active: { type: integer, minimum: 0 }
            severities:
              type: object
              additionalProperties: { type: integer, minimum: 0 }
...
    MonitoringAlert:
      type: object
      additionalProperties: false
      required: [id, deviceId, deviceName, ruleId, title, severity, state, acknowledged, openedAt, updatedAt, note]
      properties:
        id: { type: string }
        deviceId: { type: string }
        deviceName: { type: string }
        ruleId: { type: string }
        title: { type: string }
        severity: { type: string }
        state: { type: string }
        acknowledged: { type: boolean }
        openedAt: { type: [string, 'null'], format: date-time }
        updatedAt: { type: [string, 'null'], format: date-time }
        note: { type: string }
    MonitoringAlertList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/MonitoringAlert' }
    MonitoringAlertHistory:
      type: object
      additionalProperties: false
      required: [id, deviceId, deviceName, ruleId, title, severity, state, occurredAt, description]
      properties:
        id: { type: string }
        deviceId: { type: string }
        deviceName: { type: string }
        ruleId: { type: string }
        title: { type: string, maxLength: 500 }
        severity: { type: string }
        state: { type: string }
        occurredAt: { type: [string, 'null'], format: date-time }
        description: { type: string, maxLength: 4000 }
    MonitoringAlertHistoryPage:
      type: object
      additionalProperties: false
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/MonitoringAlertHistory' }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 200 }
        total: { type: integer, minimum: 0 }
...
    TimeSeriesFrame:
      type: object
      required: [id, title, unit, sampledAt, sampleMode, series]
      properties:
        id: { type: string }
        title: { type: string }
        unit: { type: string }
        sampledAt: { type: string, format: date-time }
        sampleMode: { type: string }
        pointMode: { type: string, enum: [bounded, all-available] }
        pointCount: { type: integer, minimum: 0 }
        series:
          type: array
          items:
            type: object
            required: [id, name, points]
            properties:
              id: { type: string }
              name: { type: string }
              points:
                type: array
                items:
                  type: array
                  prefixItems: [{ type: integer }, { type: [number, 'null'] }]
                  minItems: 2
                  maxItems: 2